pyproject.toml
[project]
name = "onboard_repository"
version = "0.1.0"
description = "Onboard repository logic"
readme = "README.md"
requires-python = "==3.14.3"
dependencies = [
    "pydantic==2.12.5",
    "pygithub==2.9.0",
]

[dependency-groups]
dev = [
    "pyright==1.1.408",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.1.0",
    "ruff==0.15.8",
]

[tool.ruff]
lint.select = ["ALL"]
lint.ignore = ["D", "S101"]
line-length = 120
